Refactor python part to publish on pypi + automate pypi publication
