+++ test/pdal_ign_macro/test_mark_points_to_use_for_digital_models_with_new_dimension.py
@@ -3,7 +3,9 @@ import tempfile
 import numpy as np
 import pdal
 
-from scripts.mark_points_to_use_for_digital_models_with_new_dimension import main
+from pdal_ign_macro.mark_points_to_use_for_digital_models_with_new_dimension import (
+    mark_points_to_use_for_digital_models_with_new_dimension,
+)
 
 
 def test_main():
@@ -11,7 +13,9 @@ def test_main():
     dsm_dimension = "dsm_marker"
     dtm_dimension = "dtm_marker"
     with tempfile.NamedTemporaryFile(suffix="_mark_points_output.las") as las_output:
-        main(ini_las, las_output.name, dsm_dimension, dtm_dimension, "", "")
+        mark_points_to_use_for_digital_models_with_new_dimension(
+            ini_las, las_output.name, dsm_dimension, dtm_dimension, "", ""
+        )
         pipeline = pdal.Pipeline()
         pipeline |= pdal.Reader.las(las_output.name)
         assert dsm_dimension in pipeline.quickinfo["readers.las"]["dimensions"].split(", ")
